+++ pysys/utils/threadpool.py
@@ -31,6 +31,7 @@ import queue as Queue
 
 from pysys import log
 from pysys.process.user import ProcessUser
+from pysys.internal.initlogging import pysysLogHandler, stdoutHandler
 
 # exceptions
 class NoResultsPending(Exception):
@@ -85,14 +86,14 @@ class WorkerThread(threading.Thread):
 		"""Start running the worker thread."""
 		try:
 			while True:
-				if self._dismissed.is_set() or ProcessUser.isInterruptTerminationInProgress is True:
+				if self._dismissed.is_set() or ProcessUser.isRunnerAborting is True:
 					break
 				try:
 					request = self._requests_queue.get(True, self._poll_timeout)
 				except Queue.Empty:
 					continue
 				else:
-					if self._dismissed.is_set() or ProcessUser.isInterruptTerminationInProgress is True:
+					if self._dismissed.is_set() or ProcessUser.isRunnerAborting is True: 
 						self._requests_queue.put(request)
 						break
 					try:
@@ -172,7 +173,10 @@ class ThreadPool(object):
 		self.createWorkers(num_workers, poll_timeout)
 
 	def onWorkerTerminated(self):
-		if ProcessUser.isInterruptTerminationInProgress is True: log.debug("PySys worker thread terminated after isInterruptTerminationInProgress")
+		if ProcessUser.isRunnerAborting is True: 
+			pysysLogHandler.setLogHandlersForCurrentThread([stdoutHandler])
+			log.debug("PySys worker thread terminated after isRunnerAborting")
+
 		with self.__lock:
 			self.workersRemaining -= 1