+++ test/correctness/PySys_internal_121/run.py
@@ -28,11 +28,12 @@ logFileContentsDefaultExcludes
 getExprFromFile
 pysys
 threadPoolMaxWorkers
-isInterruptTerminationInProgress
-isInterruptTerminationInProgressEvent
-isInterruptTerminationInProgressHandle
+isRunnerAborting
+isRunnerAbortingEvent
+isRunnerAbortingHandle
 grepWarnIfLineLongerThan
 grepTruncateIfLineLongerThan
+handleRunnerAbort
 		""".strip().split('\n')
 
 		def filter_member(m):
@@ -67,7 +68,7 @@ grepTruncateIfLineLongerThan
 				if m: attr.add(m.group(2))
 				m = re.search('- ``self.([a-zA-Z0-9_]+)`` ', l) # this is how we doc ivar's currently
 				if m: attr.add(m.group(1))
-		self.write_text('basetest.doc.txt', '\n'.join(sorted(a for a in attr if not a.startswith('isInterruptTerminationInProgress') ))+'\n\n'+'\n'.join(sorted(members)))
+		self.write_text('basetest.doc.txt', '\n'.join(sorted(a for a in attr if not a.startswith('isRunnerAborting') ))+'\n\n'+'\n'.join(sorted(members)))
 		
 	def validate(self):
 		self.assertGrep('basetest.doc.txt', expr='.') # check it's not empty